Rearrange provider library dependencies
providers/libcommon.a was set to be weakly depending on
providers/libfips.a and providers/libnonfips.a.  This doesn't make
sense, as libcommon is meant to be common for all.
We change this so providers/libfips.a and providers/libnonfips.a
depend on providers/libcommon.a.

It turned out that crypto/modes code was unnecessarily added to
providers/libfips.a when it should have been added to
providers/libcommon.a.  This is also changed.