Make SSL_OP_SINGLE_ECDH_USE the default and mandatory
As a follow on from the previous commit addressing the Raccoon attack
(CVE-2020-1968). We make SSL_OP_SINGLE_ECDH_USE the default and always
on (it is not possible to turn it off). This ensures that ephemeral
ECDH secrets are never reused across multiple TLS connections. While
Raccoon only affects DH, we make this change to ECDH as a precautionary
measure.

Reviewed-by: Tomas Mraz <tmraz@fedoraproject.org>