Fix rehash on the certs/demo directory
A rehash must be done on the certs/demo directory before running the test
suite, otherwise a failure will occur in test_verify. The rehashing uses
the openssl command line app to work, but it was not explicitly setting
the OPENSSL_CONF environment variable. If an explicit OPENSSL_CONF has
been set by the user, or if an openssl.cnf file happens to already exist
in the default location then that config file will be used. If that config
file is invalid (for example because it is intended for use in OpenSSL 3.0
and therefore has config modules that 1.0.2 lacks) then the openssl command
fails and the rehashing does not occur. Unfortunately the failure is silent
resulting in intermittent failures of the test suite that depend on the
environment of the user.

Reviewed-by: Paul Dale <pauli@openssl.org>
Reviewed-by: Tomas Mraz <tomas@openssl.org>