confirm bl is large enough prior to using it as an array index
in make_kn, we have this code:
k1[bl - 1] ^= bl == 16 ? 0x87 : 0x1b;

Which on newer gcc versions, triggers a stringop-overflow error.  It
occurs because gcc can't determine a definitive value range for the
passed in parameter bl.  In the case where its zero, k1[bl-1] will index
outside the array bounds

We know that won't happen, because bl is always the cipher block size,
which is non-zero.  However, given that this triggers a warning, it
breaks building with --strict-warnings

Fix is easy, just add a check in the conditional preceding this code to
confirm that we only dereference the array if bl is at lesat 1.  It
wont change functionality at all, but it quiets the warning

Addresses: https://github.openssl.org/openssl/premium/issues/31

Reviewed-by: Matt Caswell <matt@openssl.org>
Reviewed-by: Tomas Mraz <tomas@openssl.org>