Modify ssl3_cbc_digest_record to accept proper header pointer
ssl3_cbc_digest_record has in its function signature a header parameter
listed as unsigned char header[13].  This makes sense for tls
connections where the header is that size, but for ssl3 connections, the
header passed in is 75 bytes.  The pointer math all works out properly,
but this flags a gcc warning on bounds checking, thinking we're
exceeding the size of the passed buffer.  Correct it by converting the
function prototype to be unsinged char* instead of a fixed arrah size

Addresses: https://github.openssl.org/openssl/premium/issues/31

Reviewed-by: Matt Caswell <matt@openssl.org>
Reviewed-by: Tomas Mraz <tomas@openssl.org>